+++ src/ForWebforms.Tests/Bridge/Convention.cs
@@ -1,0 +1,7 @@
+﻿
+namespace ExitStrategy.ForWebforms.Tests.Bridge
+{
+    public class Convention : BasicConvention
+    {
+    }
+}
