+++ src/ForWebforms.Tests/Bridge/MvcBridgeTests.cs
@@ -9,7 +9,7 @@ using Moq;
 using Shouldly;
 using ExitStrategy.ForWebforms.Bridge;
 
-namespace ExitStrategy.ForWebforms.Tests
+namespace ExitStrategy.ForWebforms.Tests.Bridge
 {
     public class MvcBridgeTests
     {
