+++ src/ForWebforms.Tests/Controls/Convention.cs
@@ -1,0 +1,7 @@
+﻿
+namespace ExitStrategy.ForWebforms.Tests.Controls
+{
+    public class Convention : BasicConvention
+    {
+    }
+}
