+++ src/ForWebforms.Tests/ModelBinding/Convention.cs
@@ -1,0 +1,11 @@
+﻿
+namespace ExitStrategy.ForWebforms.Tests.ModelBinding
+{
+    public class Convention : BasicConvention
+    {
+        public Convention()
+        {
+            
+        }
+    }
+}
